refactor(macaroon): move core/macaroon into internal
Very soon more things will be moved into internal/macaroon. Create this
package, and move pre-existing internal code (which just so happened to
exist in core/macaroon) into the new package
