+++ api/package_test.go
@@ -37,7 +37,6 @@ func (*ImportSuite) TestImports(c *gc.C) {
 		"core/instance",
 		"core/life",
 		"core/logger",
-		"core/macaroon",
 		"core/migration",
 		"core/model",
 		"core/network",
@@ -63,6 +62,7 @@ func (*ImportSuite) TestImports(c *gc.C) {
 		"internal/featureflag",
 		"internal/http",
 		"internal/logger",
+		"internal/macaroon",
 		"internal/proxy",
 		"internal/proxy/config",
 		"internal/rpcreflect",
