+++ apiserver/authentication/user.go
@@ -20,11 +20,11 @@ import (
 
 	apiservererrors "github.com/juju/juju/apiserver/errors"
 	corelogger "github.com/juju/juju/core/logger"
-	coremacaroon "github.com/juju/juju/core/macaroon"
 	coreuser "github.com/juju/juju/core/user"
 	usererrors "github.com/juju/juju/domain/access/errors"
 	"github.com/juju/juju/internal/auth"
 	internallogger "github.com/juju/juju/internal/logger"
+	internalmacaroon "github.com/juju/juju/internal/macaroon"
 	"github.com/juju/juju/state"
 )
 
@@ -171,7 +171,7 @@ func (u *LocalUserAuthenticator) authenticateMacaroons(ctx context.Context, user
 		return nil, errors.Trace(apiservererrors.ErrUnauthorized)
 	}
 	loginMac := macaroonAuthInfo.Macaroons[index]
-	declared := checkers.InferDeclared(coremacaroon.MacaroonNamespace, loginMac)
+	declared := checkers.InferDeclared(internalmacaroon.MacaroonNamespace, loginMac)
 	username := declared[usernameKey]
 
 	// If the userTag id is not the same as the username, then the user is not
