+++ cmd/containeragent/initialize/package_test.go
@@ -56,7 +56,6 @@ func (*importSuite) TestImports(c *gc.C) {
 		"core/lease",
 		"core/life",
 		"core/logger",
-		"core/macaroon",
 		"core/machinelock",
 		"core/migration",
 		"core/model",
@@ -95,6 +94,7 @@ func (*importSuite) TestImports(c *gc.C) {
 		"internal/featureflag",
 		"internal/http",
 		"internal/logger",
+		"internal/macaroon",
 		"internal/mongo",
 		"internal/network",
 		"internal/network/debinterfaces",
